moved files to directories and perfected pathways
