+++ src/components/Directory.js
@@ -1,2 +1,2 @@
 import React, { Component } from 'react';
-import './Directory.scss';
+import '../SCSS/_Directory.scss';
